Make all hardware look like wifi-only, even if it's really a stingray
Modify wingray build so it always passes the kernel cmdline
args to force it to appear as a wifi only unit, even if it actually
has a modem present.  With this change, when the wingray build
is booted on a stingray, it'll look like a wingray.

This should be reverted before ship as devices can correctly detect
which version they are.

Change-Id: Icedddefc25a50a0663cd10948afa670e74bb8336
Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>