Fix issue #4167854: DialogWhenLarge no longer shows as a dialog on stingray
Remove the filter for only xlarge screen sizes.  Smaller screen sizes
match xlarge, so this removes resources that could possibly be needed.
For this case, DialogWhenLarge changes value based on a -large resource.
With everything except xlarge being stripped, this breaks it, causing
it to only use the base normal screen resource.

We should look at making aapt smarter about this kind of stripping,
not removing any resources that could be of use for a specific config
is specified (based on whether there are other resources that also
match and will make it inaccessible).  That needs more thought, though.

Change-Id: I92453d2c7f472174276bd744ea29727479af1aa1