Fix issue 3485235: Popping sound in video chat
The problem is due to occasional underruns in the rx path.
This is because when AEC is on, writes to pcm out driver are
driven by reads from pcm in driver: there is one write for each read
and no way for the rx path to catch up if delayed.

The fix consists in pre loading the pcm out buffers with 20ms of silence
when switching AEC on to give more marging. The problem is not completely fixed
but likelyhood for underruns and audible pops is greatly reduced.

Change-Id: I8f1de962ca6193d164448e3af5ef8cf7e4dbc308